+++ src/control/GalleryHubController.php
@@ -2,23 +2,86 @@
 
 namespace ilateral\SilverStripe\Gallery\Control;
 
-use SilverStripe\ORM\PaginatedList;
 use PageController;
+use SilverStripe\Assets\Image;
+use SilverStripe\ORM\ArrayList;
+use SilverStripe\View\ArrayData;
+use SilverStripe\ORM\PaginatedList;
 
 class GalleryHubController extends PageController
 {
-
-    public function PaginatedChildren()
+    /**
+     * Get a custom list of children with resized gallery images
+     * 
+     * @return PaginatedList
+     */
+    public function PaginatedGalleries()
     {
-        $list = $this->AllChildren();
+        $children = $this->AllChildren();
         $limit = $this->ThumbnailsPerPage;
+        $list = ArrayList::create();
 
-        $pages = PaginatedList::create(
-            $list,
-            $this->getRequest()
-        );
-        $pages->setpageLength($limit);
+        foreach ($children as $child) {
+            $image = $child->SortedImages()->first();
+            $child_data = $child->toMap();
+            $child_data["Link"] = $child->Link();
+            if ($image) {
+                $child_data["GalleryThumbnail"] = $this->ScaledImage($image, true);
+            } else {
+                $child_data["GalleryThumbnail"] = null;
+            }
+            $list->add(ArrayData::create($child_data));
+        }
+
+        $pages = PaginatedList::create($list, $this->getRequest());
+        $pages->setPageLength($limit);
 
         return $pages;
+    }
+
+    /**
+     * Generate an image based on the provided type
+     * (either )
+     *
+     * @param Image $image
+     * @param string $thumbnail generate a smaller image (based on thumbnail settings)
+     * @return void
+     */
+    protected function ScaledImage(Image $image, $thumbnail = false)
+    {
+        $img = false;
+        $background = $this->PaddedImageBackground;
+        
+        if ($thumbnail == true) {
+            $resize_type = $this->ThumbnailResizeType;
+            $width = $this->ThumbnailWidth;
+            $height = $this->ThumbnailHeight;
+        } else {
+            $resize_type = $this->ImageResizeType;
+            $width = $this->ImageWidth;
+            $height = $this->ImageHeight;
+        }
+
+        switch ($resize_type) {
+            case 'crop':
+                $img = $image->Fill($width,$height);
+                break;
+            case 'pad':
+                $img = $image->Pad($width,$height,$background);
+                break;
+            case 'ratio':
+                $img = $image->Fit($width,$height);
+                break;
+            case 'width':
+                $img = $image->ScaleWidth($width);
+                break;
+            case 'height':
+                $img = $image->ScaleHeight($height);
+                break;
+        }
+
+        $this->extend("augmentImageResize", $image, $thumbnail, $img);
+
+        return $img;
     }
 }