+++ src/model/GalleryHub.php
@@ -3,8 +3,10 @@
 namespace ilateral\SilverStripe\Gallery\Model;
 
 use Page;
+use SilverStripe\Forms\TextField;
 use SilverStripe\Forms\NumericField;
 use SilverStripe\Forms\CheckboxField;
+use SilverStripe\Forms\DropdownField;
 use ilateral\SilverStripe\Gallery\Control\GalleryHubController;
 
 /**
@@ -33,11 +35,19 @@ class GalleryHub extends Page
 
     private static $db = [
         "ShowSideBar" => "Boolean",
+        "ShowImageTitles" => "Boolean",
+        "ThumbnailWidth" => "Int",
+        "ThumbnailHeight" => "Int",
+        "ThumbnailResizeType" => "Enum(array('crop','pad','ratio','width','height'), 'crop')",
+        "PaddedImageBackground" => "Varchar",
         "ThumbnailsPerPage" => "Int"
     ];
 
     private static $defaults = [
-        "ThumbnailsPerPage" => 18
+        "ThumbnailWidth" => 150,
+        "ThumbnailHeight" => 150,
+        "ThumbnailsPerPage" => 18,
+        "PaddedImageBackground" => "ffffff"
     ];
 
     public function getControllerName() {
@@ -51,17 +61,29 @@ class GalleryHub extends Page
         $fields->addFieldsToTab(
             "Root.Settings",
             [
-                NumericField::create(
-                    'ThumbnailsPerPage',
-                    $this->fieldLabel("ThumbnailsPerPage")
-                ),
-                CheckboxField::create(
-                    'ShowSideBar',
-                    $this->fieldLabel("ShowSideBar")
-                )
+                CheckboxField::create('ShowSideBar'),
+                CheckboxField::create('ShowImageTitles'),
+                NumericField::create("ThumbnailWidth"),
+                NumericField::create("ThumbnailHeight"),
+                DropdownField::create("ThumbnailResizeType")
+                    ->setSource($this->dbObject("ThumbnailResizeType")->enumValues()),
+                NumericField::create('ThumbnailsPerPage'),
+                TextField::create("PaddedImageBackground")
             ]
         );
 
         return $fields;
+    }
+
+    public function onBeforeWrite()
+    {
+        parent::onBeforeWrite();
+
+        // default settings (if not set)
+        $defaults = $this->config()->defaults;
+        $this->ThumbnailWidth = ($this->ThumbnailWidth) ? $this->ThumbnailWidth : $defaults["ThumbnailWidth"];
+        $this->ThumbnailHeight = ($this->ThumbnailHeight) ? $this->ThumbnailHeight : $defaults["ThumbnailHeight"];
+        $this->ThumbnailsPerPage = ($this->ThumbnailsPerPage) ? $this->ThumbnailsPerPage : $defaults["ThumbnailsPerPage"];
+        $this->PaddedImageBackground = ($this->PaddedImageBackground) ? $this->PaddedImageBackground : $defaults["PaddedImageBackground"];
     }
 }